@@ -1,15 +1,20 @@
 <template>
   <div id="app">
-    <router-view></router-view>
+    <router-view v-slot="{ Component }">
+      <transition name="fade-transform" mode="out-in">
+        <component :is="Component" />
+      </transition>
+    </router-view>
   </div>
 </template>
 
-<script>
-export default {
-  name: 'app'
-}
+<script setup>
+// Vue 3 Composition API
 </script>
 
-<style rel="stylesheet/scss" lang="scss" scoped>
-
+<style lang="scss" scoped>
+#app {
+  width: 100%;
+  height: 100%;
+}
 </style>
