@@ -1,9 +1,2 @@
-import Vue from 'vue'
-import SvgIcon from '@/components/SvgIcon'// svg组件
-
-// register globally
-Vue.component('svg-icon', SvgIcon)
-
-const requireAll = requireContext => requireContext.keys().map(requireContext)
-const req = require.context('./svg', false, /\.svg$/)
-requireAll(req)
+// SVG icons are now handled by vite-plugin-svg-icons in vite.config.js
+// This file is kept for compatibility
