@@ -1,26 +1,20 @@
+import { defineStore } from 'pinia'
 import Cookies from 'js-cookie'
 
-const app = {
-  state: {
+export const useAppStore = defineStore('app', {
+  state: () => ({
     sidebar: {
       opened: !+Cookies.get('sidebarStatus')
     }
-  },
-  mutations: {
-    TOGGLE_SIDEBAR: state => {
-      if (state.sidebar.opened) {
+  }),
+  actions: {
+    toggleSideBar() {
+      if (this.sidebar.opened) {
         Cookies.set('sidebarStatus', 1)
       } else {
         Cookies.set('sidebarStatus', 0)
       }
-      state.sidebar.opened = !state.sidebar.opened
-    }
-  },
-  actions: {
-    ToggleSideBar: ({ commit }) => {
-      commit('TOGGLE_SIDEBAR')
+      this.sidebar.opened = !this.sidebar.opened
     }
   }
-}
-
-export default app
+})
