@@ -1,4 +1,4 @@
-export { default as Department } from './Department'
-export { default as Employee } from './Employee'
-export { default as Organization } from './Organization'
-export { default as Emailrel } from './Emailrel'
+export { default as Department } from './Department.vue'
+export { default as Employee } from './Employee.vue'
+export { default as Organization } from './Organization.vue'
+export { default as Emailrel } from './Emailrel.vue'
