@@ -5,9 +5,8 @@
 </template>
 
 <script>
-import echarts from 'echarts'
+import * as echarts from 'echarts'
 
-require('echarts/theme/halloween')
 
 export default {
   name: 'Tcplog',
