@@ -1,10 +1,10 @@
-export { default as Checktime } from './Checktime'
-export { default as Server } from './Server'
-export { default as Webclass } from './Webclass'
-export { default as Tcplog } from './Tcplog'
-export { default as LoginError } from './LoginError'
-export { default as Checkdep } from './Checkdep'
-export { default as Departmentcloudword } from './Departmentcloudword'
-export { default as Databaseerror } from './Databaseerror'
-export { default as Usaupload } from './Usaupload'
-export { default as Playball } from './Playball'
+export { default as Checktime } from './Checktime.vue'
+export { default as Server } from './Server.vue'
+export { default as Webclass } from './Webclass.vue'
+export { default as Tcplog } from './Tcplog.vue'
+export { default as LoginError } from './LoginError.vue'
+export { default as Checkdep } from './Checkdep.vue'
+export { default as Departmentcloudword } from './Departmentcloudword.vue'
+export { default as Databaseerror } from './Databaseerror.vue'
+export { default as Usaupload } from './Usaupload.vue'
+export { default as Playball } from './Playball.vue'
