@@ -1,15 +1,46 @@
 <template>
   <section class="app-main">
-    <transition name="fade" mode="out-in">
-      <router-view></router-view>
+    <transition name="fade-transform" mode="out-in">
+      <router-view v-slot="{ Component }">
+        <component :is="Component" :key="key" />
+      </router-view>
     </transition>
   </section>
 </template>
 
-<script>
-export default {
-  name: 'AppMain',
-  computed: {
-  }
-}
+<script setup>
+import { computed } from 'vue'
+import { useRoute } from 'vue-router'
+
+const route = useRoute()
+const key = computed(() => route.path)
 </script>
+
+<style lang="scss" scoped>
+.app-main {
+  min-height: calc(100vh - 60px);
+  width: 100%;
+  position: relative;
+  overflow: hidden;
+  padding: 20px;
+  background: rgba(255, 255, 255, 0.95);
+  backdrop-filter: blur(10px);
+  border-radius: 8px 0 0 0;
+  margin-top: 60px;
+}
+
+.fade-transform-leave-active,
+.fade-transform-enter-active {
+  transition: all 0.3s;
+}
+
+.fade-transform-enter-from {
+  opacity: 0;
+  transform: translateX(-30px);
+}
+
+.fade-transform-leave-to {
+  opacity: 0;
+  transform: translateX(30px);
+}
+</style>
